requirement.txt
asgiref==3.8.1
certifi==2024.12.14
click==8.1.8
Django==5.1.4
django-background-tasks==1.2.8
django-ckeditor==6.7.2
django-js-asset==3.0.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
mysqlclient==2.2.6
pillow==11.1.0
python-dotenv==1.0.1
redis==5.2.1
rq==2.1.0
six==1.17.0
sqlparse==0.5.3
wheel==0.45.1
